Fix compilation errors for Flutter app
- Add UserRole imports to auth and profile files
- Add instructor alias to UserRole enum for backwards compatibility
- Add quiz method implementations to CourseRepositoryImpl
- Add attemptNumber field to QuizAttemptEntity
- Add borderLight color to AppColors
- Add allLevels getter to CourseLevel enum
- Fix averageRating to ratingAverage in InstructorCourseCard
- Fix LoadCompanyDetails to use named parameter
- Add LoadCompanyMembers event to company_event.dart
- Fix score type from int to double in QuizCompleted state
- Fix LoadCourseDetails to use named parameter
- Hide EnrollInCourse from course_bloc import to resolve ambiguity

+++ lib/features/courses/presentation/bloc/student_state.dart
@@ -129,7 +129,7 @@ class QuizCompleted extends StudentState {
   final QuizAttemptEntity result;
 
   bool get passed => result.passed;
-  int get score => result.score;
+  double get score => result.score;
 
   @override
   List<Object?> get props => [result];
